Modificated logic of calculating count of iterations. More type hints fixed.

+++ src/codeplag/types.py
@@ -44,7 +44,7 @@ class FastMetrics(NamedTuple):
     operators: float
     keywords: float
     literals: float
-    weighted_average: float
+    weighted_average: np.float64
 
 
 class StructuresInfo(NamedTuple):
